pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "superagent"
version = "0.1.0"
description = "Voice-Controlled Multi-Agent Testing System for Playwright test automation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "SuperAgent Team"}
]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
    "redis>=5.0.1",
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "playwright>=1.40.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "websockets>=12.0",
    "aioredis>=2.0.1",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.22",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
superagent = "agent_system.cli:main"
kaya = "agent_system.agents.kaya:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_system*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=agent_system --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
